requirements.txt
Flask==3.1.3
bcrypt==5.0.0
beautifulsoup4==4.12.3
feedparser==6.0.12
google-auth==2.49.2
gunicorn==23.0.0
lxml==5.2.2
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
python-dotenv==1.2.2
requests==2.32.3
requests-cache==1.3.1
scipy==1.17.1
vaderSentiment==3.3.2
yfinance==1.2.0
